+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   * Add `BTPayPalVault.enablePayPalAppSwitch`
     * If set to `true` we will attempt to use the PayPal App Switch flow
 
+## 6.16.0 (2024-03-19)
+* Add `BTPayPalVaultRequest.userAuthenticationEmail` optional property
+
 ## 6.15.0 (2024-03-18)
 * [Meets Apple's new Privacy Update requirements](https://developer.apple.com/news/?id=3d8a9yyh)